move complexaggregate tests
