+++ jpa/eclipselink.jpa.test/src/it/assembly/eclipselink-advanced-model.xml
@@ -21,11 +21,9 @@
             <outputDirectory>.</outputDirectory>
             <includes>
                 <include>org/eclipse/persistence/testing/models/jpa/advanced/**</include>
-                <include>org/eclipse/persistence/testing/models/jpa/complexaggregate/**</include>
                 <include>org/eclipse/persistence/testing/models/jpa/inheritance/**</include>
                 <include>org/eclipse/persistence/testing/models/jpa/inherited/**</include>
                 <include>org/eclipse/persistence/testing/tests/jpa/advanced/**</include>
-                <include>org/eclipse/persistence/testing/tests/jpa/complexaggregate/**</include>
                 <include>org/eclipse/persistence/testing/tests/jpa/unit/**</include>
                 <include>org/eclipse/persistence/testing/tests/jpa/CMP3TestModel.class</include>
                 <include>org/eclipse/persistence/testing/tests/jpa/EntityContainerTestBase.class</include>
