+++ jpa/eclipselink.jpa.test/src/it/assembly/eclipselink-annotation-model.xml
@@ -28,7 +28,6 @@
             <outputDirectory>.</outputDirectory>
             <includes>
                 <include>org/eclipse/persistence/testing/models/jpa/advanced/**</include>
-                <include>org/eclipse/persistence/testing/models/jpa/complexaggregate/**</include>
                 <include>org/eclipse/persistence/testing/models/jpa/customer/**</include>
                 <include>org/eclipse/persistence/testing/models/jpa/inheritance/**</include>
                 <include>org/eclipse/persistence/testing/models/jpa/inherited/**</include>
