--- jpa/eclipselink.jpa.test/src/it/java/org/eclipse/persistence/testing/models/jpa/complexaggregate/Body.java
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 2011, 2020 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0,
- * or the Eclipse Distribution License v. 1.0 which is available at
- * http://www.eclipse.org/org/documents/edl-v10.php.
- *
- * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
- */
-
-// Contributors:
-//     12/17/2010-2.2 Guy Pelletier
-//       - 330755: Nested embeddables can't be used as embedded ids
-package org.eclipse.persistence.testing.models.jpa.complexaggregate;
-
-import jakarta.persistence.EmbeddedId;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name="JPA_BODY")
-public class Body {
-    @EmbeddedId
-    public Torso torso;
-
-    public Torso getTorso() {
-        return torso;
-    }
-
-    public void setTorso(Torso torso) {
-        this.torso = torso;
-    }
-
-}
