--- jpa/eclipselink.jpa.test/src/it/java/org/eclipse/persistence/testing/models/jpa/complexaggregate/Location.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) 1998, 2020 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0,
- * or the Eclipse Distribution License v. 1.0 which is available at
- * http://www.eclipse.org/org/documents/edl-v10.php.
- *
- * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
- */
-
-// Contributors:
-//     dminsky - initial API and implementation
-package org.eclipse.persistence.testing.models.jpa.complexaggregate;
-
-import jakarta.persistence.Embeddable;
-
-@Embeddable
-public class Location {
-
-    protected String address;
-    protected String countryCode;
-
-    public Location() {
-        super();
-    }
-
-    public Location(String address, String countryCode) {
-        super();
-        setAddress(address);
-        setCountryCode(countryCode);
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getCountryCode() {
-        return countryCode;
-    }
-
-    public void setCountryCode(String countryCode) {
-        this.countryCode = countryCode;
-    }
-
-}
