+++ jpa/eclipselink.jpa.test/src/it/java/org/eclipse/persistence/testing/tests/jpa/advanced/AdvancedServerTestSuite.java
@@ -21,7 +21,6 @@ import org.eclipse.persistence.testing.framework.jpa.junit.JUnitTestCase;
 import org.eclipse.persistence.testing.tests.jpa.IsolatedHashMapTest;
 import org.eclipse.persistence.testing.tests.jpa.advanced.compositepk.AdvancedCompositePKJunitTest;
 import org.eclipse.persistence.testing.tests.jpa.advanced.fetchgroup.AdvancedFetchGroupJunitTest;
-import org.eclipse.persistence.testing.tests.jpa.complexaggregate.ComplexAggregateTestSuite;
 
 /**
  * <p><b>Purpose</b>: To collect the tests that will run against Application Server only.
@@ -48,7 +47,6 @@ public class AdvancedServerTestSuite extends TestSuite {
         suite.addTest(AdvancedJunitTest.suite());
         suite.addTest(AdvancedCompositePKJunitTest.suite());
         suite.addTest(QueryCastTestSuite.suite());
-        suite.addTest(ComplexAggregateTestSuite.suiteSpring());
         if (! JUnitTestCase.isJPA10()) {
             suite.addTest(AdvancedFetchGroupJunitTest.suite());
         }
