+++ jpa/eclipselink.jpa.testapps/jpa.test.complexaggregate/src/main/java/org/eclipse/persistence/testing/models/jpa/complexaggregate/CountryDweller.java
@@ -1,0 +1,74 @@
+/*
+ * Copyright (c) 1998, 2022 Oracle and/or its affiliates. All rights reserved.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0,
+ * or the Eclipse Distribution License v. 1.0 which is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+ */
+
+// Contributors:
+//     Oracle - initial API and implementation from Oracle TopLink
+package org.eclipse.persistence.testing.models.jpa.complexaggregate;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EmbeddedId;
+import jakarta.persistence.Entity;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+
+import java.io.Serializable;
+
+@Entity
+@Table(name="CMP3_COUNTRY_DWELLER")
+public class CountryDweller implements Serializable {
+    private int age;
+    private Name name;
+    private World world;
+    private String gender;
+
+    public CountryDweller () {}
+
+    @Column(name="AGE")
+    public int getAge() {
+        return age;
+    }
+
+    @Column(name="GENDER")
+    public String getGender() {
+        return gender;
+    }
+
+    @EmbeddedId
+    public Name getName() {
+        return name;
+    }
+
+    @ManyToOne
+    public World getWorld() {
+        return world;
+    }
+
+    public void setAge(int age) {
+        this.age = age;
+    }
+
+    public void setGender(String gender) {
+        this.gender = gender;
+    }
+
+    public void setName(Name name) {
+        this.name = name;
+    }
+
+    public void setWorld(World world) {
+        this.world = world;
+    }
+
+    public String toString() {
+        return "CountryDweller: " + getName().toString();
+    }
+}
